feat: changed some things to send project

+++ umount.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-ps aux | grep '/mnt/ramdisk' | grep -v grep | awk '{print $2}' | xargs sudo kill -9
+sudo pkill -f '/mnt/ramdisk'
 
 umount /mnt/ramdisk 2>/dev/null
 
@@ -9,4 +9,6 @@ if mount | grep '/mnt/ramdisk' > /dev/null; then
     sudo umount /mnt/ramdisk
 fi
 
-rm -rf /mnt/ramdisk
+if ! mount | grep '/mnt/ramdisk' > /dev/null; then
+    rm -rf /mnt/ramdisk
+fi
